More work / menu conversion (Incomplete)
- Shop stuff
- Cleanup menu / convert old menus.
- Converted Profile menu.
- Added some E-Cash items (Golden Curse, Hats (Tentative))

+++ game/src/main/java/net/dungeonrealms/game/mastery/Utils.java
@@ -1,7 +1,7 @@
 package net.dungeonrealms.game.mastery;
 
 import net.dungeonrealms.common.Constants;
-
+import net.dungeonrealms.game.player.banks.BankMechanics;
 import org.apache.commons.io.FileUtils;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
@@ -11,6 +11,7 @@ import org.bukkit.block.Block;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.entity.Player;
+import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.inventory.meta.SkullMeta;
@@ -29,14 +30,20 @@ import java.util.function.Predicate;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
-/**
- * Created by Nick on 9/17/2015.
- */
 public class Utils {
 
     public static Logger log = Constants.log;
 
-    
+    private static SimpleDateFormat dateFormat = new SimpleDateFormat("MMM dd, YY hh:mm aa");
+
+    public static String getDateString() {
+        return getDateString(System.currentTimeMillis());
+    }
+
+    public static String getDateString(long time) {
+        return dateFormat.format(new Date(time));
+    }
+
     public static void printTrace() {
         StackTraceElement trace = new Exception().getStackTrace()[2];
 
@@ -44,7 +51,7 @@ public class Utils {
         Constants.log.info("[Database] Method: " + trace.getMethodName());
         Constants.log.info("[Database] Line: " + trace.getLineNumber());
     }
-    
+
     public static Location getLocation(String loc) {
         if (loc == null || !loc.contains(",")) return null;
         String[] args = loc.split(",");
@@ -111,6 +118,29 @@ public class Utils {
         }
         return true;
     }
+    public static boolean hasItem(Player player, Material material) {
+        return containsItem(BankMechanics.getStorage(player).inv, material) || containsItem(player.getInventory(), material);
+    }
+
+    public static boolean hasItem(Player player, String displayName) {
+        return containsItem(BankMechanics.getStorage(player).inv, displayName) || containsItem(player.getInventory(), displayName);
+    }
+
+    public static boolean containsItem(Inventory inv, Material material) {
+        return inv.contains(material);
+    }
+
+    public static boolean containsItem(Inventory inv, String itemName) {
+        itemName = ChatColor.stripColor(itemName);
+        for (ItemStack itemStack : inv.getContents()) {
+            if (itemStack == null) continue;
+            if (itemStack.getItemMeta() == null) continue;
+            if (!itemStack.getItemMeta().hasDisplayName()) continue;
+            if (ChatColor.stripColor(itemStack.getItemMeta().getDisplayName()).equalsIgnoreCase(itemName)) return true;
+        }
+
+        return false;
+    }
 
     /**
      * Given a name in the format lowercaseprefix# (e.g. us0), converts to proper format (e.g. US-0)
@@ -119,7 +149,7 @@ public class Utils {
      * @return
      */
     public static String getFormattedShardName(String name) {
-        if(name.equals("test"))return "Development";
+        if (name.equals("test")) return "Development";
         return name.split("(?=[0-9])", 2)[0].toUpperCase() + "-" + name.split("(?=[0-9])", 2)[1];
     }
 
@@ -135,7 +165,7 @@ public class Utils {
     public static int getRandomFromTier(int tier, String lvlRange) {
         int lowBase = (tier - 1) * 10;
         int highBase = tier == 5 ? 90 : lowBase;
-        
+
         int low = lvlRange.equalsIgnoreCase("high") ? lowBase + 5 : Math.max(lowBase, 1);
         int high = lvlRange.equalsIgnoreCase("low") ? highBase + 5 : highBase + 10;
         return Utils.randInt(low + 2, high + 2);
@@ -176,9 +206,9 @@ public class Utils {
     }
 
     public static String capitalize(String s) {
-    	return ucfirst(s);
+        return ucfirst(s);
     }
-    
+
     public static String ucfirst(String string) {
         return Character.toUpperCase(string.charAt(0)) + string.substring(1).toLowerCase();
     }
@@ -235,51 +265,56 @@ public class Utils {
         }
     }
 
+    public static String getDate() {
+        return getDate(System.currentTimeMillis());
+    }
+
     public static String getDate(Long milliseconds) {
         SimpleDateFormat dateFormat = new SimpleDateFormat("dd/MM/yyyy HH:mm:ss");
         dateFormat.setTimeZone(TimeZone.getTimeZone("UTC-0"));
         return dateFormat.format(new Date(milliseconds));
     }
-    
+
     /**
      * Sanitizes user input so it can be used as a file path.
      * Works by changing all non alphanumeric characters to underscore.
      * Multiple characters in a row will be treated as a single underscore.
      */
     public static String sanitizeFileName(String fileName) {
-    	return fileName.replaceAll("[^a-zA-Z0-9\\._]+", "_");
+        return fileName.replaceAll("[^a-zA-Z0-9\\._]+", "_");
     }
-    
+
     public static void removeFile(File file) {
-    	try {
-			FileUtils.forceDelete(file);
-		} catch (Exception e) {
-			e.printStackTrace();
-			Bukkit.getLogger().warning("Failed to delete " + file.getName());
-		}
+        if (!file.exists()) return;
+        try {
+            FileUtils.forceDelete(file);
+        } catch (Exception e) {
+            e.printStackTrace();
+            Bukkit.getLogger().warning("Failed to delete " + file.getName());
+        }
     }
-    
+
     /**
      * Force deletes folders / files that meet certain parameters.
      */
     public static void removeFiles(File root, Predicate<? super File> cb) {
-    	Arrays.stream(root.listFiles()).filter(cb).forEach(Utils::removeFile);
-    }
-    
-    public static String capitalizeWords(String sentence) {
-    	String formatted = "";
-		for (String s : sentence.split(" "))
-			formatted += " " + Utils.capitalize(s);
-		return formatted.length() > 1 ? formatted.substring(1) : formatted;
+        Arrays.stream(root.listFiles()).filter(cb).forEach(Utils::removeFile);
     }
 
-	public static String getItemName(ItemStack item) {
-		if (item == null || item.getType() == Material.AIR || !item.hasItemMeta())
-			return "NOTHING";
-		ItemMeta meta = item.getItemMeta();
-		if (meta.hasDisplayName())
-			return meta.getDisplayName();
-		
-		return capitalizeWords(item.getType().name().toLowerCase().replaceAll("_", " "));
-	}
+    public static String capitalizeWords(String sentence) {
+        String formatted = "";
+        for (String s : sentence.split(" "))
+            formatted += " " + Utils.capitalize(s);
+        return formatted.length() > 1 ? formatted.substring(1) : formatted;
+    }
+
+    public static String getItemName(ItemStack item) {
+        if (item == null || item.getType() == Material.AIR || !item.hasItemMeta())
+            return "NOTHING";
+        ItemMeta meta = item.getItemMeta();
+        if (meta.hasDisplayName())
+            return meta.getDisplayName();
+
+        return capitalizeWords(item.getType().name().toLowerCase().replaceAll("_", " "));
+    }
 }
